chore(startWith): move startWith TS def to startWith.d.ts
