+++ packages/core/type-definitions/combinator/startWith.d.ts
